+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 0.4.0 (2020-11-02)
 
 * Support Rails API-only mode (@janko)